Time: 0 ms (100%), Space: 13.7 MB (39.32%) - LeetHub

+++ 0224-basic-calculator/0224-basic-calculator.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     int calculate(string s) {
         stack<pair<int,int>> st;
-        long long res=0,sign=1;
+        int res=0,sign=1;
         long long num=0;
         for(char ch:s){
             if(isdigit(ch)){
